Work on examples, add travis.
