+++ .travis.yml
@@ -1,7 +1,3 @@
 language: node_js
 node_js:
-  - "0.10"
-script: "gulp coverage"
-after_success:
-  - npm install -g codeclimate-test-reporter
-  - codeclimate < coverage/lcov.info
+  - "6"
