+++ examples/coap/README.md
@@ -1,0 +1,1 @@
+Coap support is a work in progress.
