+++ lib/Grow.js
@@ -6,7 +6,6 @@ const http = require('http');
 const url = require('url');
 const coap = require('coap');
 
-
 /**
  * Grow is an extension of a Thing (which is basically a fancy event emitter).
  * See [Thing.js](https://github.com/CommonGarden/Thing.js) for more info on Things...
@@ -98,7 +97,7 @@ class Grow extends Thing {
     };
 
     /**
-     * Emits event to Grow-IoT server.
+     * Emits event to Grow-IoT server. Adds a timestamp to the event
      * @param {Object}  event  The event to emit
      * @return  this
      */
@@ -111,7 +110,7 @@ class Grow extends Thing {
 
       else if (typeof event === 'string') {
         event = {
-          message: event,
+          event,
           timestamp: new Date()
         }
       }
@@ -213,6 +212,7 @@ class Grow extends Thing {
 
   /**
    * Create a CoAP server (Needs work... see: https://github.com/mcollina/node-coap)
+   * Consider this NOT SECURE, don't use in production.
    */
   coap (options) {
     var self = this;
@@ -267,18 +267,6 @@ class Grow extends Thing {
     // the default CoAP port is 5683
     server.listen();
   }
-
-  coapClient () {
-    // make this an argument
-    var req   = coap.request('coap://localhost/jake');
-
-    req.on('response', function(res) {
-      console.log(res);
-      // res.pipe(process.stdout)
-    });
-
-    req.end();
-  }
 };
 
-export default Grow;
+module.exports = Grow;
